types: remove set and get methods from BoundRef
Once our BoundRefs start requiring an inference context to access their
data, we won't be able to call .set() and .get() on them individually.
Remove these methods, and instead add them on Context.

Doing this means that everywhere we currently call .get and .set, we
need a context available. To achieve this, we add the context to Type,
and swap the fmt::Debug/fmt::Display impls for Type and Bound so that
Type is the primary one (since it has the context object available).

The change to use BoundRef in the finalization code means we can change
our occurs-check from directly using Arc::<Bound>::as_ptr to using a
more "principled" OccursCheckId object yielded from the BoundRef. This
in turn means that we no longer need to use Arc<Bound> anywhere, and
can instead directly use Bound (which is cheap to clone and doesn't need
to be wrapped in an Arc, except when we are using Arc to obtain a
pointer-id for use in the occurs check).

Converting Arc<Bound> to Bound in turn lets us remove a bunch of
Arc::new and Arc::clone calls throughout.

Again, "API only", but there's a lot going on here.