f make Bound private, use Type as non-recursive public type
After the previous commit, we still had a memory leak in the case of
cyclic types, but for a very dumb reason: the Type structure was
carrying a Context Arc, and Type was recursive.

Instead, introduce a new private TypeInner type which is recursive and
does not have direct access to the Context. Make Bound private as well,
since it is mutually recursive with Context.

I am not thrilled with this commit and may revisit it, but it fixes the
leak and lets me get on with fuzzing before I go to bed.