types: refactor to pull some Bound constructors into Context
We have the Bound::sum and Bound::product constructors which check
whether the proposed children of the bound are complete, and if so,
construct a complete bound. If not, they construct an incomplete one.

This check involves calling Bound::final_data which locks the context
and then inspects the bound. This limits our freedom of movement to call
these constructors because we need to make sure that the context isn't
already locked. Better to pull this logic into Context where it can keep
track of locking.