types: use `Type` rather than `Bound` in the error enum
We want to make the `Bound` type private to the types module, because it
is conceptually an implementation detail of the type inferenece
algorithm. It has a weird mutual recursion with `Type` that shouldn't be
exposed in the public API.

It will also become a problem in a later commit when we change it to be
mutually recursive with a new `TypeInner` type instead. For a couple reasons:

1. `TypeInner` is a private type, so if `Bound` contains it, then
   `Bound` would need to hide its internals which is impossible for an
   enum. So we would need a `BoundInner` in addition to `TypeInner`,
   which would be a real pain to understand.

2. Neither `TypeInner` nor `Bound` will hold a reference to an inference
   context. This means that the API of `Bound` will be severely limited
   since it will no longer have the ability to introspect itself more
   than one level deep. This makes for an awkward API that should not be
   public.

This type moves a bunch of text from the main module doccomment and
makes it a non-doccomment. It is a move-only of this text; it doesn't
change anything. The existing text is somewhat out-of-date but I don't
want to change it until we're done with everything else. (It's not too
bad; the main thing is that it's now missing the `Context` object.)

I believe this is the only "controversial" commit in this PR, because
the objects in the Error type actually are, conceptually speaking,
individual Bounds rather than Types. But we are stuck between "users
should not see or care about bounds" and "the error enum should contain
all relevant context" and I think this is the best choice. We can file
an issue to revisit this maybe, though I'm willing to live with it.