simplicity-sys: update simplicity to 952e126256dd31d7ed8f4b33b99ae8cd1edabe89
This brings us to the commit immediately before the separation of
witness data in the encoding of programs.

My original goal was to bump staging one commit at a time (or several
commits in cases where there were no meaningful changes to the C code).
However, I could not build commits until this one, due to an incorrect
Tardis dependency set in libsimplicity. This one fixes the dep and is
the first one that I can run GenRustJets on.

There are three meaningful changes between our previous version of
libsecp and this one:

* The introduction of the simplicity_computeCmr function, which we don't
  need from Rust and I therefore won't bother adding bindings for;
* The addition of the txid field to the rawTransaction structure, which
  required changes to our c_set_rawTransaction shim; and
* more generally, a change to the default elements Sighash algorithm,
  which caused several Merkle roots to change in the unit tests. Some
  of these matched CMRs in the upstream unit tests, so you can check
  that my updated values again match upstream. Or you can just trust
  me that there's a good reason for these CMRs to have changed and
  that the new values are correct.