simplicity-sys: update libsimplicity to 1b85dc31d80d36dc012755e4369aeeac815476a6
This separates out the witness data from the program data. This commit
updates simplicity-sys and libsimplicity, and gets all the tests
passing, but does not update the main library for the new serialization.
Therefore the existing tests pass but the c_rust_merkle fuzztest will not.

Also fixes the binding to closeBitstream in the FFI test harness, which
I had forgotten to do back in #102.

The next commit updates the main library.