bit decoding: check that bit iterators are fully consumed
In the C library, bitstream.c, we always check that bitstreams are fully
consumed -- that there are no trailing bytes and that the final byte has
only zero bits remaining.

We were not checking this in the Rust code. Impressively, when I added
these checks, no unit tests broke except for one "maximal_sharing" which
appears to have come from the fuzzer.