types: replace Context::bind with Context::bind_product
As part of our project to make `Bound` private, we need to drop the
`Context::bind` method, which takes an arbitrary bound and applies it to
a type. This method doesn't really need to be public in so much
generality, because (for example) binding a type to a free bound is a
no-op that should never be done.

In fact, it turns out that in Simplicity's typing rules we only ever
bind to product bounds. (In some future iteration of Simplicity with
e.g. a copair combinator, we would bind to sums; but it's hard to
imagine a combinator that would require binding to units or other
complete types, since this would represent a combinator whose inputs
needed to have exact types, which is not really in the spirit of
Simplicity.)

So we can get away with just exposing a bind_product method, which takes
`Type`s and only constructs a Bound internally.