types: introduce TypeInner, make Type the only public type
Previously we had two types, `Type` and `Bound`, which were mutually
recursive. `Type` was also exposed in the public API as a type
representing an incomplete Simplicity type. `Bound` was exposed in the
public API for a couple of reasons which were eliminated over the last
few commits.

However, because `Type` is both public and unusable without a live type
inference context, this means that it needs to contain an `Arc<Context>`.
But this means that if we have a self-referential type, this leaks to a
somewhat-convoluted reference cycle in which a `Type` holds an
`Arc<Context>` which owns a set of `Bound`s which contain the original
`Type`. If you are skeptical of this, try running the test harness in
valgrind before and after this commit -- you will see that previously we
were leaking memory in the `memory_leak` unit test but now we are not.

This commit introduces a new private type `TypeInner` which is mutually
recursive with (the now-private) `Bound`. Then `Type` contains both a
`TypeInner` and an `Arc<Context>` and is not recursive, eliminating the
possibility of cyclic references. As a side-effect, it greatly reduces
the amount of Arc-cloning that is happening, which should improve
performance.