benches: move supporting functionality into library
It's a pain to do `cargo criterion --no-run` to build everything, which
is what's necessary for stuff that lives only in the benches/ directory.
So pull it into a new simplicity_bench library which can be tested using
normal commands.

It turns one pub(crate) trait into pub, which I needed to get things to
compile, but otherwise is a move-only change.