benches: check that all jets are actually benchmarked
This panics *after* all the benchmarks are done, which isn't great. But
you can speed it up when adding new jets by sticking a `continue;`
statement in each loop before the actual benchmarking happens.

As of this commit, this will cause the benchmarking code to panic at the
end since there are a bunch of new jets that aren't benchmarked. We will
add these benchmarks over the next couple commits.