benches: refactor tx environment sampling
The types around the tx sampler are a bit of a mess. We will want to add
the ability to specify that all the outputs should be fees, which is
hard to do with the current arrangements. Use a builder pattern instead
which will let us tack arbitrary new conditions onto the transactions.