benches: refactor value handling and sampling logic for core jets
Our existing logic has fairly heavy setup costs because it constructs
entire recursive simplicity::Value structures, only to bit-serialize
then destroy them. I suspect this is bleeding into some of the low-cost
jet measurements due to issues with criterion (though after this commit,
the benchmark numbers are largely the same, so perhaps I was wrong).

Furthermore, and more importantly, our sampling logic does not really do
what we want. Our original goal was to sample the "90% percentile" time
for each jet, i.e. an upper bound on the time a jet takes for "90%" of
its inputs. criterion is not really capable of that, so instead we ran
100 benchmarks for every jet and then took the worst average of those.

However, within each benchmark we mixed all sorts of different input
types in ad-hoc ways. For uniform bit inputs this is fine, but for
inputs like group elements, we really want to be benchmarking on-curve
and off-curve points separately so that we can determine which one is
actually the worst-case and specifically benchmark that.

The new logic introduces an `InputSample` trait which is capable of
separating out and composing distributions, giving us an exact count of
how many different distributions (so, e.g. for jets that just take
random bits, we have only a couple, while for something like gej_add we
have many, including adding the same point to itself). This means that
the number of distributions, rather than being a fixed 100, now ranges
from 1 for the unit-taking jets up to 400 for the linear verify and
point verify jets. The result is a much faster, and yet more useful,
benchmark run.