feat: Create {256,512}-bit value infallibly
If we create a Value from &[u8; 32] then we assert the length of the
slice at compile time and cannot fail at runtime. Similar for &[u8; 64].
This is similar to rust-bitcoin's from_byte_array with the difference
that Value takes the bytes by reference.

It turns out, in rust-simplicity we can simply convert every call of
{u256, u512}_from_slice to a call of {u256, u512} without problem.
In the jets-bench subcrate, we have to convert [u32; 8] to [u8; 32]
or &[u8; 33] to &[u8; 32] via try_into. I think this is still worth
doing, and we can see in the code that these conversions will never fail.

Keep {u256, u512}_from_slice for backwards compatability.