update libsimplicity to 121b2951d8cca3d73dccd6fcbe721f2195025067
This update renames a bunch of symbols from x to simplicity_x. This
commit has a small number of non-mechanical changes to address that.

It also updates the benchmarks and therefore the CMRs and AMRs of many
jets.

The one "judgement call" was to update wrapper.h in accordance with
https://github.com/BlockstreamResearch/simplicity/pull/248#issuecomment-2248830161
to cover all the jet wrapper renames.