refactor: Convert Cost to Weight
Cost is a weight in milli weight units. Sometimes we need to work in
weight units. This commit updates the API of Cost so that Cost can be
converted to bitcoin::Weight and vice versa. Function inputs or outputs
that are weight units are converted to bitcoin::Weight to show the unit
of measurement.

Because bitcoin::Weight works in terms of u64 while Cost works in terms
of u32, this commit adds U32Weight, which is a private newtype that
safely translates between u64 and u32.

Because we use bitcoin::Weight in elements contexts, we require the
bitcoin feature for elements code. We will be able to use
elements::Weight once this struct is added to rust-elements.