feat: Zero value
It is useful to populate witness nodes that will be pruned out after the
first run of the Bit Machine. We don't need to ask the user to provide
made-up witness data for branches that are not executed.

To this end, we can compute a zero value for each Simplicity type. The
zero value serializes to a string of zeroes, and it is easy to compute.