feat: Finalize witness nodes with(out) pruning
We construct redeem programs by finalizing witness programs. We often
want the finalized program to be pruned, so we introduce the
finalize_pruned method. This is method that most people should use.

In cases where the transaction environment is not yet known, there is
the finalize_unpruned method. We could make this method private in the
future, depending on how we want to handle pruning in our API. We might
need to come up with a nicer way to track which program is pruned for
which environment, and so on.