refactor: Remove Error::DisconnectCommitTime
This error was thrown when a disconnected branch is populated during
construction of a CommitNode. The error was intended to inform users
that the disconnected branches are thrown away, but this information is
not useful. It is a general fact that CommitNode omits disconnected
branches. We don't need to throw a runtime error in case a disconnected
branch exists.

Removing Error::DisconnectedCommitTime eliminates a useless runtime
error path.