fix: Handle missing assembly fragments
The policy satisfier needs to produce a Simplicity program that matches
a given CMR. When an assembly fragment is missing, then we can use
hidden nodes to represent it. Hidden<N> simulates this behavior by
propagating hidden nodes upwards towards the next case node. If the root
node becomes hidden, then the entire program is unsatisfiable.