refactor: Simplify policy satisfier
Use hidden nodes to signify unsatisfiability of sub-expressions.
When a sub-expression is unsatisfiable, then it makes no sense to
execute it. We can just as well keep its CMR in a hidden node.
This simplifies the code and it solves a potential problem of the
previous design: a hidden node could be marked as satisfiable.