value: replace as_left/as_right/as_product with to_* variants
Currently, because we have a recursive `Value` type, we can return
sub-values by reference. However, if we move to a non-recursive type
this will no longer be possible, since sub-values will need to be newly
constructed objects.

In advance of changing the internals of `ValueInner`, change these
methods.

Because the internals of `Value` are just `Arc`s, it's cheap to clone
them so there is not much efficiency cost to doing this.