bit machine: check resource bounds on construction
Currently it is possible to ask the bit machine to do arbitrarily large
allocations, and to panic the program by overflowing when computing the
size of its allocations.

This is not the place to do budget checks, but we can at least do some
sanity checks. Here we insist that programs not have any types that
exceed 2 gigabits (256 megabytes), which will at least prevent
explosions. We can revisit it when somebody shows up with a usecase for
massive programs.