ci: Run unit tests on Mac OS
We have a job that runs all unit tests for all feature combinations on
Linux. I find it excessive to run all of this on Mac OS, so I added a
separate job to run tests with all features enabled (which is 1 feature
combination instead of all combinations).

In a follow-up PR, we can add Windows to the feature matrix of this CI
job, but with the current code base this job would fail. We need to
spend time to debug this.