simplicity and simplicity-sys: bump major versions
In another PR I add a regression test against the released 0.3.0 version
of rust-simplicity. It breaks some tooling to have two 0.3.0s in play at
once -- one from this repo and one from crates.io.

Take this opportunity to bump versions to 0.4.0 to avoid this conflict.

Once we get this into a reasonable bug-free state we should do another
release.