rename `FirstPassIhr` to `Imr`
In a recent PR we renamed IMR (Identity Merkle Root) to the IHR
(Identity Hash Root). We did this via mass-rename which turned the
`FirstPassImr` object into `FirstPassIhr`.

The correct terminogy is now that `FirstPassImr` is simply `Imr`. This
object continues to be a mostly-internal-only object which is used in
service of computing the `Ihr`, which is what users are likely to
interact with.