clippy: address all current lints
Mostly they are complaining about the new "format specifiers can
directly hold the variable, if that variable is a single token, and if
the format specifier is in println or format but not write or assert, so
you should convert 40% of your instances to this different format,
because it makes a big diff and leaves your code in an unreadable state"
lint, which personally I don't feel is a net benefit.

Copy the whole list of whitelisted clippy lints from rust-elements.