clippy: enable lossless-cast lint
There is a huge list of lints I'd like to enable (currently in
rust-elements and a few other crates). As a start, enable the
cast_lossless one, which I think is very important: it complaints if you
use a cast where usize::from() would suffice.

The benefit of the from() syntax is that it will compile iff the cast
cannot truncate. So the reader of the code can tell immediately that
there is no danger of changing the value, and if the lint is on,
conversely the reader can tell that anything that *is* a cast is
something that needs further scrutiny.