update libsimplicity to cad9a63292ea8fdd5c1615e0322986f6f996d42f
Here I did the bare minimum renaming in the rest of the library to get
things to compile. This was in three files:

* in simplicity-sys/build.rs I had to tweak some paths
* In simplicity-sys/depend/jets_wrapper.c I had to tweak a path
* In simplicity-sys/depend/env.c tweaked some paths, added typedefs, and
  updated the signature of c_set_txEnv based on the same function in env.c
  from Haskell/cbits in libsimpliciity.

THEN in order for the tests to pass and not segfault, I had to implement
three API changes in simplicity-sys/src/tests/ffi.rs (and update the call
sites in simplicity-sys/src/tests/mod.rs):

* simplicity_decodeMallocDag now takes a callback
* simplicity_mallocTypeInference now takes a callback (and defining this
  callback required adding bindings for a bunch of aux types, annoyingly)
* SimplicityErr adds a new OVERWEIGHT variant (and extends the unit tests
  to hit this)

Rust being extremely agressive about UB in unsafe code, by dropping any
of the three of these you can get very weird-looking segfaults (e.g. an
immediate jump to a null address with no backtrace, for the missing enum
variant but a unit test that creates in from C code).

Anyway, I tried to minimize this diff but wasn't totally successful. The
next commits do the "rest" of the upgrade work, which consists of a bunch
of mechanical renaming.