simplicity-sys: add links= field to the Cargo.toml file
This field tells Cargo what library we're linking in, and helps it
provide sane error messages (rather than linker spew) when somebody
tries to compile multiple versions of the library that link the
same thing.

When we provide a links field we should also provide a build field.
This is set to build.rs by default, but crate2nix seems to get confused
when one is present but not the other.