node: add generic byte encoding
Right now we have encoding methods on RedeemNode and CommitNode. They are
not generic because RedeemNode has witness data while CommitNode does not.

This is a problem because it means that we cannot use the bit encoding to
implement generic fmt::Display impls. Instead we iterate and implement
fmt::Display in terms of a dump of each of the nodes in the program, with
newlines and everything, which is not really useful as a Display impl.

It's also an API annoyance; it means that we have to re-implement the
encoding method on every single node type individualy; there is no way
to encode "just the program" for redeem nodes, there are no encoding
functions for ConstructNode or NamedConstructNode even though it's
totally possible to encode these by calling encode_program directly, etc.
(I think it was a semi-deliberate design choice not to foreground the
encoding of ConstructNodes. But this is dumb. It's no less confusing
to encode ConstructNodes than to encode CommitNodes; in both cases it
is impossible to share "correctly" but there is still a natural way to
do so.)

While we are at it, stop taking a BitWriter and just take a generic Write.
There is no circumstance where the user should be bit-encoding a program
into the middle of another bitstream, since we split the witness data off
from the program data. (In fact, arguably there is no circumstance where
a user needs to work with bitwriters *at all*.) We weren't even using it
nicely as a BitWriter, since we'd always call `flush` after writing. Just
acknowledge that we will always write a whole number of bytes by taking
an io::Write.

Also while we're at it, return the count of program and witness bits
separately when encoding to a generic writer. We were returning the sum,
which since the split between witness and program, is a basically useless
quantity. (Not that we were using the return value of this function
anywhere..)

f deal with the "sharing id" of namedcommitnode being the name