human_encoding: make "sharing ID" match the sharing ID of the underlying node
In NamedNode<Construct> and NamedNode<Commit> we had set the sharing ID to
be the name of the node. The idea here was that if two nodes had different
names in the human-readable encoding, even if they were identical, that
they would not be shared since they were "conceptually different".

This idea was at best confusing, and at worst dangerous. "Sharing" is a
property of the bit encoding, not the human-readable encoding, and
pretending otherwise is just wrong.

I'm not actually sure what I intended by making the sharing ID be the name.
Notably, in the bit encoding we didn't actually use the sharing ID! We
override it in NamedCommitNode::encode, where we first convert the named
commit node to a normal commit node before encoding it. So AFAICT this
confused notion of sharing ID was never actually used. Some evidence for
this is that this commit doesn't break any unit tests..

However, in the next commit I will make encoding generic, and I won't be
able to use the "convert to a commitnode" hack anymore.

Anyway this is all moot since the human-readable encoding has been almost
completely superceded by SimplicityHL.