node: add from_str method to Construct, Commit and RedeemNode
We cannot use the std from_str trait because these all construct an Arc<Self>
rather than a Self. And we cannot implement this generically on Node because
the rules for decoding are different for each node type (we have progressively
tighter correctness requirements, and with RedeemNode we need extra witness
data).