error: split up top-level error type into DecodeError and FinalizeError
Essentially, DecodeError is used when constructing a program from the
bit encoding and FinalizeError is used when constructing a program using
the library.

Along the way, tighten up some error returns for functions that return
e.g. crate::Error when they actually only ever yield type errors.

We drop the Error::Policy variant which was never used. The old Error
enum now has a single variant, which annoyingly is used by the
Haskell-generated code so we can't remove it without roundtripping
through upstream. Will try to do this before the next release but it's
not a huge priority.

In the next commit we'll add a ParseError which wraps DecodeError and
adds some string-encoding-related variants.