add ghost-cell dependency
This one is coauthored by Ralf Jung and has a paper (which I read) and a
formal proof (which I did not). It would be hard to find a more trustworthy
dependency.

Also, we have a clear need for this -- currently we have Mutexes all over
our type checker code which are both slow and hard to reason about.