types: move innards of Context into a ContextInner type
The Context object contains an Arc, which is probably the right design
decision since these are supposed to be "handles" to a type inference
context which can be cheaply cloned everywhere. Without this there
would be a ton of Arc::new and Arc::clone noise throughout the codebase.

But we want to make this structure more elaborate, which requires putting
more stuff in the Arc, which requires a new struct. So do that in its
own commit, to try to reduce the amount of noise in the next one.