node: thread invariant 'brand lifetime through freeking everything
All this commit does is patch a 'brand lifetime (the name comes from the
GhostCell paper) into `types::Context`. In the next commit this lifetime
will be associated with a GhostToken but for now it's unused.

However, putting an invariant lifetime onto `types::Context` means that
we need an invariant lifetime on `Type`, which means one on `ConstructNode`
and the `*Constructible` traits, and on any functions which create these
things.

In fact, this is exactly what we want! All these types and functions are
ones that should be tied to a single type-inference context, and by use
of GhostCells we will get compile-time assurance that users are not trying
to mix types from one program with types from another. As a second benefit,
this should get rid of the mutex mess in the typechecker and union-bound
code.

IOW this is an extremely large commit, but it doesn't do anything at all
except add lifetime annotations, and they're all good.