types: introduce 'Incomplete' type to represent incomplete type bounds
Right now our types::Error enum has a couple variants that include bare
`Type`s in them which are intended to represent invalid types. It isn't
good hygeine to do this, because these Types carry with them information
about the type context (which is supposed to be an ephemeral thing
created during parsing and destroyed before parsing completes).

It also limits our ability to provide useful representation of incomplete
types in error messages, because the bare Bound type does not contain any
data about why a bound was invalid. The user is also forced to be careful
with these invalid objects, especially in the case of an occurs-check
failure, when any attempt to iterate on the type will infinite-loop.

This commit does the bare minimum to define a new type for use in error
messages. In a followup we probably want to clean things up a bit -- at
the very least we should try to construct some sort of rich "occurs check
failure" information. We may also want to define an IncompleteArrow,
which would make the code in human_encoding a bit tidier.

The attentive reviewer may notice calls to `lock.drop()` inserted before
calls to Incomplete::from_bound_ref. Without these calls, the code deadlocks.
There is no compiler help or anything to keep track of this. This kind of
bullshit is why I'm so excited to use GhostCell, where the compiler (and
borrowck) will do all the work for us.

Finally, in the next commit when we add a 'brand lifetime marker to Type,
we will be forced to avoid saving this type in our Error enum, because
then Type will be unable to outlive its type context (and errors must
outlive the context, since the context is ephemeral and the error is
something we return to the user).