human_encoding: pass context object as argument to Node::to_witness_node
Same justification as before. If you are constructing a ConstructNode (or
a WitnessNode, which is basically the same thing), you should keep explicit
track of your context objects.

Also, prior to this commit, if you were converting named nodes to witness
nodes for the purpose of combining them, you could not do this because
to_witness_node was making a fresh type context for each conversion.