types: use GhostToken for type context identity
Prior to this commit we were using pointer values and assertions to
check that users were using type contexts consistently (i.e. not
using different contexts for the same program). This worked reasonably
well, but was (a) slow, (b) revealed errors only at runtime, and (c)
provided almost no help in guiding the user to using the correct
context in the correct places.

Instead, we can use a ghost_cell::GhostToken as an identity, which is
a special type which can be used to assign a unique invariant "brand"
lifetime to every type context. Then everything that holds a context
reference, i.e. every type bound and ConstructNode, inherits this
brand, and the compiler will ensure that you don't mix brands *and*
that you don't fail to mix brands. And it will do it all using the
borrowchecker, meaning that your code won't compile if you're wrong,
but the markers get erased prior to codegen so that the compilation
time and code size aren't meaningfully increased.

Furthermore, the use of invariant lifetimes ensures that we catch
mistakes in both directions: if the user uses two contexts where they
should've used one, that's an error (which our existing assertions
will probably catch); *and* if the user uses one context where they
should've used two, that's probably also an error.

The tradeoff here is that you can no longer construct context objects
directly and then pass them around. You need to construct them using
Context::with_context, which provides a fresh context object to a
closure, and the context won't outlive the closure. (This is how we
catch "if you use one context where you should've used two" errors;
to make this mistake and still have your code compile, you need to
put two program constructions into the same closure, which hopefully
will be visibly wrong.)

So far, based on my updating of our unit tests and all our code, this
limitation doesn't seem to be a big deal. It does require a bunch of
reindentation but it hasn't made the code any harder to write or
structure -- and where it did force restructuring (e.g. in the Satisfy
trait in the previous commit) it was because the original code was
actually wrong.

Another tradeoff is that when you get stuff wrong the compiler errors
are pretty rough. But I think we'll have to live with this, and
hopefully future compilers will learn to understand GhostCell and
GhostToken constructions better.

Definitely want to review this using
    git show --color-moved-ws=ignore-all-space
because the bulk of the diff is just me re-indenting unit tests to
wrap them in Context::with_context.